# http-servers/express-server/swagger-homework/api/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: NodeJS Mentoring Programm
  version: 1.0.0
  title: Homework8
  contact:
    email: Yauheni_Rohau@epam.com
host: localhost:10010
basePath: /api
...
schemes:
- http
paths:
  /products:
    get:
      tags:
      - products
      summary: Get all products
      description: Get all products
      operationId: getAllProducts
      produces:
      - application/json
      parameters: []
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Product'
    post:
      tags:
      - products
      summary: Add product
      description: Add product
      operationId: addProduct
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: data for product
        required: false
        schema:
          $ref: '#/definitions/Product'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Product'
    x-swagger-router-controller: products
  /products/{id}:
    get:
      tags:
      - products
      summary: Get product with id
      description: Get product with id
      operationId: getProductWithId
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: Product id
        required: true
        type: string
      responses:
        200:
          description: results
          schema:
            $ref: '#/definitions/Product'
        400:
          description: Error
          schema:
            type: object
    delete:
      tags:
      - products
      summary: Delete product by id
      description: Delete product by id
      operationId: deleteProductById
      parameters:
      - name: id
        in: path
        description: Product id
        required: true
        type: string
      responses:
        200:
          description: OK
          schema:
            type: object
        400:
          description: Error
          schema:
            type: object
    x-swagger-router-controller: products
...
definitions:
  Product:
    type: object
    required:
    - name
    properties:
      name:
        type: string
      type:
        type: string
        minLength: 3
      price:
        type: number
      exist:
        type: boolean
        default: false
    example:
      name: name
      type: type
      price: 100
      exist: true
...